changed db connection from  sqlite to mysql

+++ Data/ChatDbContext.cs
@@ -12,7 +12,7 @@ using VueChatApp.Features.QrCode.Entities;
 
 namespace VueChatApp.Data
 {
-    public class ChatDbContext : IdentityDbContext<SystemUser, IdentityRole<int>, int>
+    public class ChatDbContext : IdentityDbContext<SystemUser, IdentityRole<long>, long>
     {
         public ChatDbContext(DbContextOptions<ChatDbContext> options) : base(options) => Database.EnsureCreated();
         public DbSet<ChatMessage> ChatMessages { get; set; }
