+++ Features/Chat/Models/OutgoingChatMessageViewModel.cs
@@ -7,12 +7,12 @@ namespace VueChatApp.Features.Chat.Models
 {
     public class OutgoingChatMessageViewModel
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Message { get; set; }
         public string Type { get; set; }
-        public int FromId { get; set; }
+        public long FromId { get; set; }
         public String FromUserName { get; set; }
-        public int ToId { get; set; }
+        public long ToId { get; set; }
         public String ToUserName { get; set; }
         public DateTime? CreatedAt { get; set; }
         public DateTime? UpdatedAt { get; set; }
